refactor: consolidate build, dev, and start commands into index.ts and update version to 0.0.7
