+++ packages/@buntal/cli/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@buntal/cli",
   "description": "Command line interface for Buntal",
-  "version": "0.0.6",
+  "version": "0.0.7",
   "module": "index.ts",
   "type": "module",
   "homepage": "https://buntaljs.org",
